Move extensions to more appropriate files. Cleanup Manifest file.
