+++ app/src/main/java/com/andrew/liashuk/phasediagram/common/Events.kt
@@ -1,7 +1,5 @@
 package com.andrew.liashuk.phasediagram.common
 
-import kotlinx.coroutines.channels.Channel
-
 abstract class Event
 
 class ShowProgress : Event()
@@ -9,18 +7,3 @@ class ShowProgress : Event()
 class HideProgress : Event()
 
 class ShowToast(val message: String) : Event()
-
-fun Channel<Event>.showProgress(show: Boolean) {
-    trySend(if (show) ShowProgress() else HideProgress())
-}
-
-inline fun <T> Channel<Event>.withProgress(block: () -> T) : T {
-    showProgress(true)
-    return block().also {
-        showProgress(false)
-    }
-}
-
-fun Channel<Event>.showToast(message: String) {
-    trySend(ShowToast(message))
-}
