+++ app/src/main/java/com/andrew/liashuk/phasediagram/viewmodal/DiagramViewModel.kt
@@ -7,13 +7,13 @@ import com.andrew.liashuk.phasediagram.R
 import com.andrew.liashuk.phasediagram.common.DispatcherProvider
 import com.andrew.liashuk.phasediagram.common.Event
 import com.andrew.liashuk.phasediagram.common.ResourceResolver
-import com.andrew.liashuk.phasediagram.common.showProgress
-import com.andrew.liashuk.phasediagram.common.showToast
-import com.andrew.liashuk.phasediagram.common.withProgress
 import com.andrew.liashuk.phasediagram.ext.firstNotNull
 import com.andrew.liashuk.phasediagram.ext.getMutableStateFlow
 import com.andrew.liashuk.phasediagram.ext.isEmpty
 import com.andrew.liashuk.phasediagram.ext.runCoroutine
+import com.andrew.liashuk.phasediagram.ext.showProgress
+import com.andrew.liashuk.phasediagram.ext.showToast
+import com.andrew.liashuk.phasediagram.ext.withProgress
 import com.andrew.liashuk.phasediagram.logic.DiagramCalculator
 import com.andrew.liashuk.phasediagram.types.PhaseData
 import com.github.mikephil.charting.data.Entry
